Allocates a pointer

+++ 0x0C-more_malloc_free/100-realloc.c
@@ -36,7 +36,7 @@ void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 
 	for (loop = 0; loop <= (int)old_size; loop++)
 	{
-		if (!(new_size < loop))
+		if (!((int)new_size < loop))
 		temp[loop] = ((char *)ptr)[loop];
 	}
